refactor: 중복 파일 삭제 및 파일 이동으로 인한 패키지 수정
