+++ app/src/main/java/woowacourse/omok/domain/RowType.kt
@@ -1,4 +1,4 @@
-package omok.domain
+package woowacourse.omok.domain
 
 enum class RowType(val value: Int) {
     A(0),
