+++ app/src/main/java/woowacourse/omok/domain/Rule.kt
@@ -1,4 +1,4 @@
-package omok.domain
+package woowacourse.omok.domain
 
 interface Rule {
     fun isInvalid(
