+++ app/src/main/java/woowacourse/omok/domain/Stone.kt
@@ -1,3 +1,3 @@
-package omok.domain
+package woowacourse.omok.domain
 
 data class Stone(val position: Position, val color: StoneType)
