+++ app/src/main/java/woowacourse/omok/domain/StoneType.kt
@@ -1,4 +1,4 @@
-package omok.domain
+package woowacourse.omok.domain
 
 enum class StoneType {
     BLACK,
