+++ app/src/main/java/woowacourse/omok/library/MoreThanFiveRule.kt
@@ -1,6 +1,6 @@
 package omok.library
 
-import omok.domain.StoneType
+import woowacourse.omok.domain.StoneType
 
 class MoreThanFiveRule(boardSize: Int) : OmokRule(boardSize) {
     override fun validate(
