+++ app/src/main/java/woowacourse/omok/library/OmokRule.kt
@@ -1,6 +1,6 @@
 package omok.library
 
-import omok.domain.StoneType
+import woowacourse.omok.domain.StoneType
 
 abstract class OmokRule(
     boardSize: Int,
