--- src/main/kotlin/omok/Main.kt
@@ -1,12 +1,0 @@
-package omok
-
-import omok.controller.OmokGame
-import omok.view.InputView
-import omok.view.OutputView
-
-fun main() {
-    val inputView = InputView()
-    val outputView = OutputView()
-    val omokGame = OmokGame(inputView, outputView)
-    omokGame.start()
-}
