--- src/main/kotlin/omok/domain/Board.kt
@@ -1,34 +1,0 @@
-package omok.domain
-
-class Board(private val rule: Rule) {
-    val stones: Stones = Stones()
-    val grid: Array<Array<StoneType>> = Array(BOARD_SIZE) { Array(BOARD_SIZE) { StoneType.EMPTY } }
-
-    fun put(
-        position: Position,
-        color: StoneType,
-    ): Stone {
-        val stone = Stone(position, color)
-        isValidPosition(stone)
-        if (grid[position.row][position.column] != StoneType.EMPTY) {
-            throw IllegalArgumentException(ERROR_STONE_ALREADY_PLACED)
-        }
-        stones.add(stone)
-        grid[position.row][position.column] = stone.color
-        return stone
-    }
-
-    fun isFull(): Boolean = grid.all { row -> row.all { column -> column != StoneType.EMPTY } }
-
-    private fun isValidPosition(stone: Stone) {
-        if (stone.color == StoneType.BLACK && rule.isInvalid(stones, stone, grid)) {
-            throw IllegalArgumentException(ERROR_INVALID_BLACK_STONE)
-        }
-    }
-
-    companion object {
-        const val BOARD_SIZE = 15
-        private const val ERROR_STONE_ALREADY_PLACED = "이미 돌이 놓여진 위치입니다. 다시 입력해주세요."
-        private const val ERROR_INVALID_BLACK_STONE = "흑돌이 놓을 수 없는 금수 위치입니다. 다시 입력해주세요."
-    }
-}
