--- src/main/kotlin/omok/domain/RowType.kt
@@ -1,19 +1,0 @@
-package omok.domain
-
-enum class RowType(val value: Int) {
-    A(0),
-    B(1),
-    C(2),
-    D(3),
-    E(4),
-    F(5),
-    G(6),
-    H(7),
-    I(8),
-    J(9),
-    K(10),
-    L(11),
-    M(12),
-    N(13),
-    O(14),
-}
