--- src/main/kotlin/omok/domain/Rule.kt
@@ -1,9 +1,0 @@
-package omok.domain
-
-interface Rule {
-    fun isInvalid(
-        stones: Stones,
-        lastPlacedStone: Stone,
-        playingBoard: Array<Array<StoneType>>,
-    ): Boolean
-}
