--- src/main/kotlin/omok/domain/Stone.kt
@@ -1,3 +1,0 @@
-package omok.domain
-
-data class Stone(val position: Position, val color: StoneType)
