--- src/main/kotlin/omok/domain/StoneType.kt
@@ -1,7 +1,0 @@
-package omok.domain
-
-enum class StoneType {
-    BLACK,
-    WHITE,
-    EMPTY,
-}
