--- src/main/kotlin/omok/domain/Turn.kt
@@ -1,17 +1,0 @@
-package omok.domain
-
-class Turn {
-    var color: StoneType = StoneType.BLACK
-        private set
-
-    fun next() {
-        color =
-            if (color == StoneType.BLACK) {
-                StoneType.WHITE
-            } else {
-                StoneType.BLACK
-            }
-    }
-
-    fun isWhite(): Boolean = color == StoneType.WHITE
-}
