--- src/test/kotlin/omok/Fixture.kt
@@ -1,326 +1,0 @@
-package omok
-
-import omok.domain.Position
-import omok.domain.Stone
-import omok.domain.StoneType
-
-val Black = StoneType.BLACK
-val White = StoneType.WHITE
-
-val A1 = Position(0, 14)
-val A2 = Position(0, 13)
-val A3 = Position(0, 12)
-val A4 = Position(0, 11)
-val A5 = Position(0, 10)
-val A6 = Position(0, 9)
-val A7 = Position(0, 8)
-val A8 = Position(0, 7)
-val A9 = Position(0, 6)
-val A10 = Position(0, 5)
-val A11 = Position(0, 4)
-val A12 = Position(0, 3)
-val A13 = Position(0, 2)
-val A14 = Position(0, 1)
-val A15 = Position(0, 0)
-
-val B1 = Position(1, 14)
-val B2 = Position(1, 13)
-val B3 = Position(1, 12)
-val B4 = Position(1, 11)
-val B5 = Position(1, 10)
-val B6 = Position(1, 9)
-val B7 = Position(1, 8)
-val B8 = Position(1, 7)
-val B9 = Position(1, 6)
-val B10 = Position(1, 5)
-val B11 = Position(1, 4)
-val B12 = Position(1, 3)
-val B13 = Position(1, 2)
-val B14 = Position(1, 1)
-val B15 = Position(1, 0)
-
-val C1 = Position(2, 14)
-val C2 = Position(2, 13)
-val C3 = Position(2, 12)
-val C4 = Position(2, 11)
-val C5 = Position(2, 10)
-val C6 = Position(2, 9)
-val C7 = Position(2, 8)
-val C8 = Position(2, 7)
-val C9 = Position(2, 6)
-val C10 = Position(2, 5)
-val C11 = Position(2, 4)
-val C12 = Position(2, 3)
-val C13 = Position(2, 2)
-val C14 = Position(2, 1)
-val C15 = Position(2, 0)
-
-val D1 = Position(3, 14)
-val D2 = Position(3, 13)
-val D3 = Position(3, 12)
-val D4 = Position(3, 11)
-val D5 = Position(3, 10)
-val D6 = Position(3, 9)
-val D7 = Position(3, 8)
-val D8 = Position(3, 7)
-val D9 = Position(3, 6)
-val D10 = Position(3, 5)
-val D11 = Position(3, 4)
-val D12 = Position(3, 3)
-val D13 = Position(3, 2)
-val D14 = Position(3, 1)
-val D15 = Position(3, 0)
-
-val E1 = Position(4, 14)
-val E2 = Position(4, 13)
-val E3 = Position(4, 12)
-val E4 = Position(4, 11)
-val E5 = Position(4, 10)
-val E6 = Position(4, 9)
-val E7 = Position(4, 8)
-val E8 = Position(4, 7)
-val E9 = Position(4, 6)
-val E10 = Position(4, 5)
-val E11 = Position(4, 4)
-val E12 = Position(4, 3)
-val E13 = Position(4, 2)
-val E14 = Position(4, 1)
-val E15 = Position(4, 0)
-
-val F1 = Position(5, 14)
-val F2 = Position(5, 13)
-val F3 = Position(5, 12)
-val F4 = Position(5, 11)
-val F5 = Position(5, 10)
-val F6 = Position(5, 9)
-val F7 = Position(5, 8)
-val F8 = Position(5, 7)
-val F9 = Position(5, 6)
-val F10 = Position(5, 5)
-val F11 = Position(5, 4)
-val F12 = Position(5, 3)
-val F13 = Position(5, 2)
-val F14 = Position(5, 1)
-val F15 = Position(5, 0)
-
-val G1 = Position(6, 14)
-val G2 = Position(6, 13)
-val G3 = Position(6, 12)
-val G4 = Position(6, 11)
-val G5 = Position(6, 10)
-val G6 = Position(6, 9)
-val G7 = Position(6, 8)
-val G8 = Position(6, 7)
-val G9 = Position(6, 6)
-val G10 = Position(6, 5)
-val G11 = Position(6, 4)
-val G12 = Position(6, 3)
-val G13 = Position(6, 2)
-val G14 = Position(6, 1)
-val G15 = Position(6, 0)
-
-val H1 = Position(7, 14)
-val H2 = Position(7, 13)
-val H3 = Position(7, 12)
-val H4 = Position(7, 11)
-val H5 = Position(7, 10)
-val H6 = Position(7, 9)
-val H7 = Position(7, 8)
-val H8 = Position(7, 7)
-val H9 = Position(7, 6)
-val H10 = Position(7, 5)
-val H11 = Position(7, 4)
-val H12 = Position(7, 3)
-val H13 = Position(7, 2)
-val H14 = Position(7, 1)
-val H15 = Position(7, 0)
-
-val I1 = Position(8, 14)
-val I2 = Position(8, 13)
-val I3 = Position(8, 12)
-val I4 = Position(8, 11)
-val I5 = Position(8, 10)
-val I6 = Position(8, 9)
-val I7 = Position(8, 8)
-val I8 = Position(8, 7)
-val I9 = Position(8, 6)
-val I10 = Position(8, 5)
-val I11 = Position(8, 4)
-val I12 = Position(8, 3)
-val I13 = Position(8, 2)
-val I14 = Position(8, 1)
-val I15 = Position(8, 0)
-
-val J1 = Position(9, 14)
-val J2 = Position(9, 13)
-val J3 = Position(9, 12)
-val J4 = Position(9, 11)
-val J5 = Position(9, 10)
-val J6 = Position(9, 9)
-val J7 = Position(9, 8)
-val J8 = Position(9, 7)
-val J9 = Position(9, 6)
-val J10 = Position(9, 5)
-val J11 = Position(9, 4)
-val J12 = Position(9, 3)
-val J13 = Position(9, 2)
-val J14 = Position(9, 1)
-val J15 = Position(9, 0)
-
-val K1 = Position(10, 14)
-val K2 = Position(10, 13)
-val K3 = Position(10, 12)
-val K4 = Position(10, 11)
-val K5 = Position(10, 10)
-val K6 = Position(10, 9)
-val K7 = Position(10, 8)
-val K8 = Position(10, 7)
-val K9 = Position(10, 6)
-val K10 = Position(10, 5)
-val K11 = Position(10, 4)
-val K12 = Position(10, 3)
-val K13 = Position(10, 2)
-val K14 = Position(10, 1)
-val K15 = Position(10, 0)
-
-val L1 = Position(11, 14)
-val L2 = Position(11, 13)
-val L3 = Position(11, 12)
-val L4 = Position(11, 11)
-val L5 = Position(11, 10)
-val L6 = Position(11, 9)
-val L7 = Position(11, 8)
-val L8 = Position(11, 7)
-val L9 = Position(11, 6)
-val L10 = Position(11, 5)
-val L11 = Position(11, 4)
-val L12 = Position(11, 3)
-val L13 = Position(11, 2)
-val L14 = Position(11, 1)
-val L15 = Position(11, 0)
-
-val M1 = Position(12, 14)
-val M2 = Position(12, 13)
-val M3 = Position(12, 12)
-val M4 = Position(12, 11)
-val M5 = Position(12, 10)
-val M6 = Position(12, 9)
-val M7 = Position(12, 8)
-val M8 = Position(12, 7)
-val M9 = Position(12, 6)
-val M10 = Position(12, 5)
-val M11 = Position(12, 4)
-val M12 = Position(12, 3)
-val M13 = Position(12, 2)
-val M14 = Position(12, 1)
-val M15 = Position(12, 0)
-
-val N1 = Position(13, 14)
-val N2 = Position(13, 13)
-val N3 = Position(13, 12)
-val N4 = Position(13, 11)
-val N5 = Position(13, 10)
-val N6 = Position(13, 9)
-val N7 = Position(13, 8)
-val N8 = Position(13, 7)
-val N9 = Position(13, 6)
-val N10 = Position(13, 5)
-val N11 = Position(13, 4)
-val N12 = Position(13, 3)
-val N13 = Position(13, 2)
-val N14 = Position(13, 1)
-val N15 = Position(13, 0)
-
-val O1 = Position(14, 14)
-val O2 = Position(14, 13)
-val O3 = Position(14, 12)
-val O4 = Position(14, 11)
-val O5 = Position(14, 10)
-val O6 = Position(14, 9)
-val O7 = Position(14, 8)
-val O8 = Position(14, 7)
-val O9 = Position(14, 6)
-val O10 = Position(14, 5)
-val O11 = Position(14, 4)
-val O12 = Position(14, 3)
-val O13 = Position(14, 2)
-val O14 = Position(14, 1)
-val O15 = Position(14, 0)
-
-val FourFour =
-    listOf(
-        Stone(C15, Black), Stone(C14, Black), Stone(C12, Black), Stone(C11, Black), Stone(C10, Black),
-        Stone(D12, Black), Stone(G12, Black), Stone(I12, Black), Stone(J12, Black),
-        Stone(J9, Black), Stone(H8, Black), Stone(J8, Black), Stone(K8, Black),
-        Stone(H7, Black), Stone(H6, Black), Stone(E6, Black), Stone(J6, Black),
-        Stone(E5, Black), Stone(F5, Black), Stone(G5, Black), Stone(G4, Black),
-    )
-
-val ThreeThree =
-    listOf(
-        Stone(D14, Black), Stone(D13, Black), Stone(C12, Black), Stone(E12, Black),
-        Stone(B6, Black), Stone(C5, Black), Stone(E6, Black), Stone(E5, Black),
-        Stone(J9, Black), Stone(M12, Black), Stone(M10, Black), Stone(N9, Black),
-        Stone(K6, Black), Stone(K3, Black), Stone(M4, Black), Stone(N4, Black),
-    )
-
-val Full =
-    listOf(
-        Stone(A15, Black), Stone(B15, Black), Stone(C15, White), Stone(D15, White),
-        Stone(E15, Black), Stone(F15, Black), Stone(G15, White), Stone(H15, White),
-        Stone(I15, Black), Stone(J15, Black), Stone(K15, White), Stone(L15, White),
-        Stone(M15, Black), Stone(N15, Black), Stone(O15, White), Stone(A14, White),
-        Stone(B14, White), Stone(C14, Black), Stone(D14, Black), Stone(E14, White),
-        Stone(F14, White), Stone(G14, Black), Stone(H14, Black), Stone(I14, White),
-        Stone(J14, White), Stone(K14, Black), Stone(L14, Black), Stone(M14, White),
-        Stone(N14, White), Stone(O14, Black), Stone(A13, Black), Stone(B13, Black),
-        Stone(C13, White), Stone(D13, White), Stone(E13, Black), Stone(F13, Black),
-        Stone(G13, White), Stone(H13, White), Stone(I13, Black), Stone(J13, Black),
-        Stone(K13, White), Stone(L13, White), Stone(M13, Black), Stone(N13, Black),
-        Stone(O13, White), Stone(A12, White), Stone(B12, White), Stone(C12, Black),
-        Stone(D12, Black), Stone(E12, White), Stone(F12, White), Stone(G12, Black),
-        Stone(H12, Black), Stone(I12, White), Stone(J12, White), Stone(K12, Black),
-        Stone(L12, Black), Stone(M12, White), Stone(N12, White), Stone(O12, Black),
-        Stone(A11, Black), Stone(B11, Black), Stone(C11, White), Stone(D11, White),
-        Stone(E11, Black), Stone(F11, Black), Stone(G11, White), Stone(H11, White),
-        Stone(I11, Black), Stone(J11, Black), Stone(K11, White), Stone(L11, White),
-        Stone(M11, Black), Stone(N11, Black), Stone(O11, White), Stone(A10, White),
-        Stone(B10, White), Stone(C10, Black), Stone(D10, Black), Stone(E10, White),
-        Stone(F10, White), Stone(G10, Black), Stone(H10, Black), Stone(I10, White),
-        Stone(J10, White), Stone(K10, Black), Stone(L10, Black), Stone(M10, White),
-        Stone(N10, White), Stone(O10, Black), Stone(A9, Black), Stone(B9, Black),
-        Stone(C9, White), Stone(D9, White), Stone(E9, Black), Stone(F9, Black),
-        Stone(G9, White), Stone(H9, White), Stone(I9, Black), Stone(J9, Black),
-        Stone(K9, White), Stone(L9, White), Stone(M9, Black), Stone(N9, Black),
-        Stone(O9, White), Stone(A8, White), Stone(B8, White), Stone(C8, Black),
-        Stone(D8, Black), Stone(E8, White), Stone(F8, White), Stone(G8, Black),
-        Stone(H8, Black), Stone(I8, White), Stone(J8, White), Stone(K8, Black),
-        Stone(L8, Black), Stone(M8, White), Stone(N8, White), Stone(O8, Black),
-        Stone(A7, Black), Stone(B7, Black), Stone(C7, White), Stone(D7, White),
-        Stone(E7, Black), Stone(F7, Black), Stone(G7, White), Stone(H7, White),
-        Stone(I7, Black), Stone(J7, Black), Stone(K7, White), Stone(L7, White),
-        Stone(M7, Black), Stone(N7, Black), Stone(O7, White), Stone(A6, White),
-        Stone(B6, White), Stone(C6, Black), Stone(D6, Black), Stone(E6, White),
-        Stone(F6, White), Stone(G6, Black), Stone(H6, Black), Stone(I6, White),
-        Stone(J6, White), Stone(K6, Black), Stone(L6, Black), Stone(M6, White),
-        Stone(N6, White), Stone(O6, Black), Stone(A5, Black), Stone(B5, Black),
-        Stone(C5, White), Stone(D5, White), Stone(E5, Black), Stone(F5, Black),
-        Stone(G5, White), Stone(H5, White), Stone(I5, Black), Stone(J5, Black),
-        Stone(K5, White), Stone(L5, White), Stone(M5, Black), Stone(N5, Black),
-        Stone(O5, White), Stone(A4, White), Stone(B4, White), Stone(C4, Black),
-        Stone(D4, Black), Stone(E4, White), Stone(F4, White), Stone(G4, Black),
-        Stone(H4, Black), Stone(I4, White), Stone(J4, White), Stone(K4, Black),
-        Stone(L4, Black), Stone(M4, White), Stone(N4, White), Stone(O4, Black),
-        Stone(A3, Black), Stone(B3, Black), Stone(C3, White), Stone(D3, White),
-        Stone(E3, Black), Stone(F3, Black), Stone(G3, White), Stone(H3, White),
-        Stone(I3, Black), Stone(J3, Black), Stone(K3, White), Stone(L3, White),
-        Stone(M3, Black), Stone(N3, Black), Stone(O3, White), Stone(A2, White),
-        Stone(B2, White), Stone(C2, Black), Stone(D2, Black), Stone(E2, White),
-        Stone(F2, White), Stone(G2, Black), Stone(H2, Black), Stone(I2, White),
-        Stone(J2, White), Stone(K2, Black), Stone(L2, Black), Stone(M2, White),
-        Stone(N2, White), Stone(O2, Black), Stone(A1, Black), Stone(B1, Black),
-        Stone(C1, White), Stone(D1, White), Stone(E1, Black), Stone(F1, Black),
-        Stone(G1, White), Stone(H1, White), Stone(I1, Black), Stone(J1, Black),
-        Stone(K1, White), Stone(L1, White), Stone(M1, Black), Stone(N1, Black),
-        Stone(O1, White),
-    )
